+++ src/gufo/thor/services/static.py
@@ -52,11 +52,7 @@ class StaticService(NocService):
         r = super().get_compose_volumes(config, svc) or []
         # Filter out settings and crashinfo
         if r:
-            r = [
-                x
-                for x in r
-                if not x.startswith(("./etc/noc/settings.yml:", "crashinfo:"))
-            ]
+            r = [x for x in r if not x.startswith("crashinfo:")]
         if config.noc.path:
             # Preserve /ui/pkg to not be overriden with repo
             r.append("/opt/noc/ui/pkg")