+++ CHANGELOG.md
@@ -10,6 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 To see unreleased changes, please see the [CHANGELOG on the master branch](https://github.com/gufolabs/gufo_thor/blob/master/CHANGELOG.md) guide.
 
+## [Upcoming]
+
+## Changed
+
+* Configs are deployed using docker's `configs` section. 
+* NOC's config are mounted in /etc/noc/settings.yml
+
 ## 0.11.0 - 2025-09-12
 
 ## Fixed